src/main/java/com/suppenstube/model/entity/News.java
package com.suppenstube.model.entity;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Table;

@Entity
@Table(name = "news")
public class News extends BaseEntity {
    @Column(nullable = false)
    private String headline;

    @Column(nullable = false)
    private String text;

    public String getHeadline() {
        return headline;
    }

    public void setHeadline(String headline) {
        this.headline = headline;
    }

    public String getText() {
        return text;
    }

    public void setText(String text) {
        this.text = text;
    }
}
